@@ -1,6 +1,6 @@
 module github.com/cashapp/spirit
 
-go 1.23.0
+go 1.24.0
 
 require (
 	github.com/alecthomas/kong v0.7.1
