@@ -1,7 +1,6 @@
 package check
 
 import (
-	"context"
 	"testing"
 
 	"github.com/cashapp/spirit/pkg/statement"
@@ -13,22 +12,22 @@ func TestIllegalClauseCheck(t *testing.T) {
 	r := Resources{
 		Statement: statement.MustNew("ALTER TABLE t1 ADD INDEX (b), ALGORITHM=INPLACE"),
 	}
-	err := illegalClauseCheck(context.Background(), r, logrus.New())
+	err := illegalClauseCheck(t.Context(), r, logrus.New())
 	assert.Error(t, err)
 	assert.ErrorContains(t, err, "contains unsupported clause")
 
 	r.Statement = statement.MustNew("ALTER TABLE t1  ADD c INT, ALGORITHM=INPLACE, LOCK=shared")
-	err = illegalClauseCheck(context.Background(), r, logrus.New())
+	err = illegalClauseCheck(t.Context(), r, logrus.New())
 	assert.Error(t, err)
 	assert.ErrorContains(t, err, "contains unsupported clause")
 
 	r.Statement = statement.MustNew("ALTER TABLE t1  ADD c INT, lock=none")
-	err = illegalClauseCheck(context.Background(), r, logrus.New())
+	err = illegalClauseCheck(t.Context(), r, logrus.New())
 	assert.Error(t, err)
 	assert.ErrorContains(t, err, "contains unsupported clause")
 
 	r.Statement = statement.MustNew("ALTER TABLE t1 engine=innodb, algorithm=copy")
-	err = illegalClauseCheck(context.Background(), r, logrus.New())
+	err = illegalClauseCheck(t.Context(), r, logrus.New())
 	assert.Error(t, err)
 	assert.ErrorContains(t, err, "contains unsupported clause")
 }
