@@ -1,7 +1,6 @@
 package check
 
 import (
-	"context"
 	"testing"
 
 	"github.com/cashapp/spirit/pkg/statement"
@@ -14,25 +13,25 @@ func TestRename(t *testing.T) {
 	r := Resources{
 		Statement: statement.MustNew("ALTER TABLE t1 RENAME TO newtablename"),
 	}
-	err := renameCheck(context.Background(), r, logrus.New())
+	err := renameCheck(t.Context(), r, logrus.New())
 	assert.Error(t, err)
 	assert.ErrorContains(t, err, "renames are not supported")
 
 	r.Statement = statement.MustNew("ALTER TABLE t1 RENAME COLUMN c1 TO c2")
-	err = renameCheck(context.Background(), r, logrus.New())
+	err = renameCheck(t.Context(), r, logrus.New())
 	assert.Error(t, err)
 	assert.ErrorContains(t, err, "renames are not supported")
 
 	r.Statement = statement.MustNew("ALTER TABLE t1 CHANGE c1 c2 VARCHAR(100)")
-	err = renameCheck(context.Background(), r, logrus.New())
+	err = renameCheck(t.Context(), r, logrus.New())
 	assert.Error(t, err)
 	assert.ErrorContains(t, err, "renames are not supported")
 
 	r.Statement = statement.MustNew("ALTER TABLE t1 CHANGE c1 c1 VARCHAR(100)") //nolint: dupword
-	err = renameCheck(context.Background(), r, logrus.New())
+	err = renameCheck(t.Context(), r, logrus.New())
 	assert.NoError(t, err)
 
 	r.Statement = statement.MustNew("ALTER TABLE t1 ADD INDEX (anothercol)")
-	err = renameCheck(context.Background(), r, logrus.New())
+	err = renameCheck(t.Context(), r, logrus.New())
 	assert.NoError(t, err) // safe modification
 }
