@@ -16,19 +16,19 @@ import (
 func TestMetadataLock(t *testing.T) {
 	lockTableInfo := table.TableInfo{SchemaName: "test", TableName: "test"}
 	logger := logrus.New()
-	mdl, err := NewMetadataLock(context.Background(), testutils.DSN(), &lockTableInfo, logger)
+	mdl, err := NewMetadataLock(t.Context(), testutils.DSN(), &lockTableInfo, logger)
 	assert.NoError(t, err)
 	assert.NotNil(t, mdl)
 
 	// Confirm a second lock cannot be acquired
-	_, err = NewMetadataLock(context.Background(), testutils.DSN(), &lockTableInfo, logger)
+	_, err = NewMetadataLock(t.Context(), testutils.DSN(), &lockTableInfo, logger)
 	assert.ErrorContains(t, err, "lock is held by another connection")
 
 	// Close the original mdl
 	assert.NoError(t, mdl.Close())
 
 	// Confirm a new lock can be acquired
-	mdl3, err := NewMetadataLock(context.Background(), testutils.DSN(), &lockTableInfo, logger)
+	mdl3, err := NewMetadataLock(t.Context(), testutils.DSN(), &lockTableInfo, logger)
 	assert.NoError(t, err)
 	assert.NoError(t, mdl3.Close())
 }
@@ -37,7 +37,7 @@ func TestMetadataLockContextCancel(t *testing.T) {
 	lockTableInfo := table.TableInfo{SchemaName: "test", TableName: "test-cancel"}
 
 	logger := logrus.New()
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(t.Context())
 	mdl, err := NewMetadataLock(ctx, testutils.DSN(), &lockTableInfo, logger)
 	assert.NoError(t, err)
 	assert.NotNil(t, mdl)
@@ -49,7 +49,7 @@ func TestMetadataLockContextCancel(t *testing.T) {
 	<-mdl.closeCh
 
 	// Confirm the lock is released by acquiring a new one
-	mdl2, err := NewMetadataLock(context.Background(), testutils.DSN(), &lockTableInfo, logger)
+	mdl2, err := NewMetadataLock(t.Context(), testutils.DSN(), &lockTableInfo, logger)
 	assert.NoError(t, err)
 	assert.NotNil(t, mdl2)
 	assert.NoError(t, mdl2.Close())
@@ -59,7 +59,7 @@ func TestMetadataLockRefresh(t *testing.T) {
 	lockTableInfo := table.TableInfo{SchemaName: "test", TableName: "test-refresh"}
 	logger := logrus.New()
 
-	mdl, err := NewMetadataLock(context.Background(), testutils.DSN(), &lockTableInfo, logger, func(mdl *MetadataLock) {
+	mdl, err := NewMetadataLock(t.Context(), testutils.DSN(), &lockTableInfo, logger, func(mdl *MetadataLock) {
 		// override the refresh interval for faster testing
 		mdl.refreshInterval = 2 * time.Second
 	})
@@ -70,7 +70,7 @@ func TestMetadataLockRefresh(t *testing.T) {
 	time.Sleep(5 * time.Second)
 
 	// Confirm the lock is still held
-	_, err = NewMetadataLock(context.Background(), testutils.DSN(), &lockTableInfo, logger)
+	_, err = NewMetadataLock(t.Context(), testutils.DSN(), &lockTableInfo, logger)
 	assert.ErrorContains(t, err, "lock is held by another connection")
 
 	// Close the lock
@@ -105,11 +105,11 @@ func TestMetadataLockLength(t *testing.T) {
 
 	logger := logrus.New()
 
-	_, err := NewMetadataLock(context.Background(), testutils.DSN(), &lockTableInfo, logger)
+	_, err := NewMetadataLock(t.Context(), testutils.DSN(), &lockTableInfo, logger)
 	// No error anymore after using a hash of the table name
 	assert.NoError(t, err)
 
-	_, err = NewMetadataLock(context.Background(), testutils.DSN(), empty, logger)
+	_, err = NewMetadataLock(t.Context(), testutils.DSN(), empty, logger)
 	assert.ErrorContains(t, err, "metadata lock table info is nil")
 }
 
@@ -129,7 +129,7 @@ func TestMetadataLockRefreshWithConnIssueSimulation(t *testing.T) {
 	logger.SetLevel(logrus.DebugLevel)
 
 	// create a new MetadataLock with a short refresh interval for testing
-	mdl, err := NewMetadataLock(context.Background(), testutils.DSN(), &lockTableInfo, logger, func(mdl *MetadataLock) {
+	mdl, err := NewMetadataLock(t.Context(), testutils.DSN(), &lockTableInfo, logger, func(mdl *MetadataLock) {
 		mdl.refreshInterval = 2 * time.Second
 	})
 	assert.NoError(t, err)
@@ -144,7 +144,7 @@ func TestMetadataLockRefreshWithConnIssueSimulation(t *testing.T) {
 	time.Sleep(4 * time.Second)
 
 	// confirm the lock is still held by attempting to acquire it with a new connection
-	_, err = NewMetadataLock(context.Background(), testutils.DSN(), &lockTableInfo, logger)
+	_, err = NewMetadataLock(t.Context(), testutils.DSN(), &lockTableInfo, logger)
 	assert.ErrorContains(t, err, "lock is held by another connection")
 
 	// close the lock
